Update install.sh to detect the OS and parse the OpenVINO version
Previously, the `install.sh` script relied on an old version of Ubuntu
for installation and, by default, installed an old version of
OpenVINO. This change allows the script to detect which Linux version is
running and uses a new default for OpenVINO--2021.4.752. The script is
still Ubuntu-specific (e.g., relies on `apt`) but that is an expected OS
for GitHub runners, the target for this script.