ci: inspect environment changes
When troubleshooting the environment setup that this action attempts to
do using OpenVINO's provided `setupvars.sh` script, it is nice to be
able to inspect the changes made to the environment and system path.
This PR is not intended to be merged, but rather as an experimental
branch for checking what is actually happening on the GitHub runners.