Use latest version with APT (#25)
According to the OpenVINO [docs], we should now be able to install OpenVINO with the `apt.sh` script. To do so, we need to use the OS release name (e.g., `ubuntu22`) instead of codename (e.g., `jammy`) when we create the `deb` file for APT. Also, we need to install slightly different prerequisites when it comes to TBB: Ubuntu 20.04 installations require `libtbb2` but all later versions, like Ubuntu 22.04 need `libtbb12`. We can verify this in the downloaded `install_dependencies/install_openvino_dependencies.sh` script.

[docs]: https://docs.openvino.ai/2023.2/openvino_docs_install_guides_installing_openvino_apt.html#installing-openvino-runtime