Inform GitHub action of the release we plan to use (#28)
This is a Linux-only hack to avoid dependency failures with TBB; the
`ubuntu20` installation wants `libtbb2` whereas the `ubuntu22`
installation wants `libtbb12`. Since either release can be installed on
either OS, we need to properly install the right dependency (as long as
we aren't using APT, which should just do the right thing).